+++ decomp/ST.exe/functions/004E3120/decomp.c
@@ -19,7 +19,7 @@ undefined4 __thiscall TLOBaseTy::sub_004E3120(TLOBaseTy *this)
   short local_a;
 
   if (*(int *)&this->field_0x361 == 2) {
-    thunk_FUN_004e5f20((int)this->field_0024,*(uint *)&this->field_0x369);
+    thunk_FUN_004e5f20(this->field_0024,*(uint *)&this->field_0x369);
     if (DAT_008117bc != (undefined4 *)0x0) {
       uVar1 = *(undefined2 *)&this->field_0024;
       memset(local_24, 0, 0x20); /* compiler bulk-zero initialization */
