+++ decomp/ST.exe/functions/004EBF10/decomp.c
@@ -7,7 +7,7 @@ void __thiscall FUN_004ebf10(void *this,int param_1)
 
   if (*(int *)((int)this + 0x508) == param_1) {
     *(undefined4 *)((int)this + 0x4d0) = 5;
-    thunk_FUN_004cbf70((int)this);
+    thunk_FUN_004cbf70(this);
     TLOBaseTy::RotateSpr(this,0);
     if (*(int *)((int)this + 0x5ac) == 0x52) {
       iVar1 = *(int *)this;
