+++ decomp/ST.exe/functions/005B69F0/decomp.c
@@ -29,8 +29,7 @@ ChatGlassTy::OutChGlProc
     iVar3 = Library::MSVCRT::__setjmp3(local_4c.jumpBuffer,0);
     puVar2 = local_8;
     if (iVar3 == 0) {
-      Library::DKW::DDX::FUN_006c68f0
-                ((int)param_1,param_4,param_5,param_6,param_7,(int)(local_8 + 1));
+      Library::DKW::DDX::FUN_006c68f0(param_1,param_4,param_5,param_6,param_7,(int)(local_8 + 1));
       FUN_006c6850((int)param_1,param_4,param_5,param_6,param_7,*puVar2);
       iVar3 = param_5 + -0x16 + param_7;
       Library::DKW::DDX::FUN_006c53b0
