+++ decomp/ST.exe/functions/005B99F0/decomp.c
@@ -28,8 +28,7 @@ MMMObjTy::OutEdProc(MMMObjTy *this,AnonShape_GLOBAL_0080759C_9638EF10 *param_1,u
     errorCode = Library::MSVCRT::__setjmp3(local_4c.jumpBuffer,0);
     puVar2 = local_8;
     if (errorCode == 0) {
-      Library::DKW::DDX::FUN_006c68f0
-                ((int)param_1,param_4,param_5,param_6,param_7,(int)(local_8 + 1));
+      Library::DKW::DDX::FUN_006c68f0(param_1,param_4,param_5,param_6,param_7,(int)(local_8 + 1));
       FUN_006c6850((int)param_1,param_4,param_5,param_6,param_7,*puVar2);
       Library::DKW::DDX::FUN_006c53b0
                 ((int)param_1,param_4,param_5 + 0x16,param_4 + -1 + param_6,param_5 + 0x16,*puVar2);
