+++ decomp/ST.exe/functions/006A0150/decomp.c
@@ -12,8 +12,8 @@ undefined4 __thiscall CGenerate::sub_006A0150(CGenerate *this)
   int iVar3;
 
   pbVar1 = (byte *)Library::Ourlib::MFSTMAP::mfTMapCreate
-                             (this->field_0228,this->field_022C,(int)this->field_0234,
-                              this->field_0239,0x20);
+                             (this->field_0228,this->field_022C,this->field_0234,this->field_0239,
+                              0x20);
   this->field_000C = pbVar1;
   piVar2 = Library::Ourlib::MFSTMAP::FUN_006f0620((short *)pbVar1);
   this->field_0008 = piVar2;
